New scroll handler
The scroll handler is no longer added to every video on the page, instead it is added to the document itself. On every scroll it checks to see if the mouse is over a video and then changes the volume of the video.

Fixed some misnamed strings on the popup script.